+++ src/main/scala/co/topl/brambl/cli/BramblCliParams.scala
@@ -21,7 +21,7 @@ object BramblCliSubCmd extends Enumeration {
 
   val invalid, init, recoverkeys, utxobyaddress, blockbyheight, blockbyid,
       transactionbyid, create, prove, broadcast, currentaddress, list, add,
-      inspect, exportvk, importvks, sync, setinteraction, balance = Value
+      inspect, exportvk, importvks, sync, setinteraction, listinteraction, balance = Value
 }
 
 sealed abstract class NetworkIdentifiers(